src/requirements.txt
insightface
onnxruntime
opencv-python
numpy
google-cloud-storage
google-cloud-pubsub
flask
flask-cors
